requirements.txt
colorama==0.4.6
keyboard==0.13.5
mouse==0.7.1
mss==9.0.1
Pillow==10.3.0
PyYAML==6.0.1
PyYAML==6.0.1
